agent-backend/requirements.txt
fastapi

uvicorn

langgraph

langchain

langchain-ollama

langchain-community

chromadb

pyautogui

pillow

ollama